move function outside `route.ts`
